ocr_server/requirements.txt

fastapi
uvicorn
paddlepaddle-gpu
paddleocr
python-multipart
opencv-python
numpy
pydantic
pydantic-settings
requests
pytest
httpx
Pillow
